chore: add importAs to metadata generated for unit tests
This is necessary so that, when generating ngInjectableDefs, ngc can
correctly determine the import location for injected dependencies.
Without this, the generated imports use a relative path to dist instead
of e.g. "@angular/cdk/overlay". In real life it's not a problem because
the packages are resolved to node_modules.